feat(30/10/2022): plasma-manager, custom.nix, modules cleaning
Rather than have laptop specific config in the configuration.nix file I
move it to the hosts folder. This will allow me to install my config on
a system without enabling xserver in the future.

A cool open source module that save our plasma config in the right
place. Themes and such are not configured right now.

NB: we have config for plasma both in a nixos module and a home-manager
module which I am not a fan of.. This will do for the time being.