feat(nvim): better daily note, automatically remove empty notes
The `<M-d>` keybind automatically create a dated note in the `~/daily`
folder. The keybind is sometime also used to navigate to older notes for
convenience, as such some empty file are created in the folder.

We now delete those notes on command launch. A note is considered empty
if it contains less than two lines of text (the first line being the
note name).